+++ changelog.md
@@ -3,6 +3,7 @@
 - Added Relocate Mission type
 - Added Vanilla stats for: Easy, normal and hard total completed missions
 - Added stats as tooltip of InfoButton
+- Added vanilla stat for missions rerolled
 
 # 0.3.2
 - Fixed some dedicated servers not saving properly the mission data